+++ src/main/java/io/bookwright/di/ApiModule.java
@@ -18,6 +18,7 @@ import io.bookwright.api.semaphore.system.SemaphoreSystemApi;
 import io.bookwright.api.semaphore.tasks.SemaphoreTasksApi;
 import io.bookwright.api.semaphore.templates.SemaphoreTemplatesApi;
 import io.bookwright.api.semaphore.users.SemaphoreUsersApi;
+import io.bookwright.api.testenvironment.runners.DynamicRunnerLauncherApi;
 import io.bookwright.config.Configs;
 import io.bookwright.config.MainConfig;
 import io.bookwright.teardown.TeardownStorage;
@@ -57,6 +58,13 @@ public class ApiModule extends AbstractModule {
   @Named("semaphore")
   Retrofit semaphoreRetrofit(MainConfig config) {
     return RetrofitFactory.create(config.apiBaseUrl());
+  }
+
+  @Provides
+  @Singleton
+  @Named("runnerFixture")
+  Retrofit runnerFixtureRetrofit(MainConfig config) {
+    return RetrofitFactory.create(config.runnerFixtureBaseUrl());
   }
 
   @Provides
@@ -161,5 +169,11 @@ public class ApiModule extends AbstractModule {
   @Singleton
   SemaphoreUsersApi semaphoreUsersApi(@Named("semaphore") Retrofit retrofit) {
     return retrofit.create(SemaphoreUsersApi.class);
+  }
+
+  @Provides
+  @Singleton
+  DynamicRunnerLauncherApi dynamicRunnerLauncherApi(@Named("runnerFixture") Retrofit retrofit) {
+    return retrofit.create(DynamicRunnerLauncherApi.class);
   }
 }